rename, move clientSocket to game/
