+++ frontend/pages/game/ClientSocket.tsx
