+++ frontend/pages/game/home/index.tsx
@@ -1,11 +1,11 @@
 import type { NextPage } from 'next';
 import { Layout } from './components/Layout';
-import { GameDisplay } from './components/Display';
+import { Display } from './components/Display';
 
 const Home: NextPage = () => {
   return (
     <Layout title="Matching">
-      <GameDisplay />
+      <Display />
     </Layout>
   );
 };
